+++ crates/rtlola-parser/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add `offset(by: off, or: dft)` syntactic sugar
 - add `->` syntactic sugar
 - add instance aggregation
+- add explicit annotations for global and local frequencies, e.g. `@Global(1Hz)` and `@Local(1Hz)`
 
 ### Changed
 - Trigger are now parsed as output streams and not handled separately.